Refactor TS and break apart a few things.
